Split up railties/test/command/application_test.rb
Follow-up to #47208.

Prior to #47208, `Rails::Command::ApplicationTest` appears to have been
specifically intended to test `Rails::Command::ApplicationCommand`.
Therefore, this commit extracts the tests that were added by #47208 out
from `railties/test/command/application_test.rb` into
`railties/test/command/help_integration_test.rb`, and moves
`application_test.rb` from `railties/test/command` to
`railties/test/commands` (with the rest of the command-specific tests).
